requirements.txt
customtkinter
psutil
py-cpuinfo
WMI
fpdf2
Pillow